requirements.txt
clip==1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
ftfy==6.3.1
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.1
pillow==10.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
regex==2024.9.11
six==1.16.0
sympy==1.13.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.66.6
typing_extensions==4.9.0
wcwidth==0.2.13
